% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVC_mle.R
\name{cov_par}
\alias{cov_par}
\title{Extact Covariance Parameters

Method to extract the covariance parameters from an \code{\link{SVC_mle}} object.}
\usage{
cov_par(object, ...)
}
\arguments{
\item{object}{\code{\link{SVC_mle}} object}

\item{...}{further arguments}
}
\value{
vector with covariance parameters
}
\description{
Extact Covariance Parameters

Method to extract the covariance parameters from an \code{\link{SVC_mle}} object.
}
