% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVC_mle.R
\name{fitted.SVC_mle}
\alias{fitted.SVC_mle}
\title{Extact Model Fitted Values

Method to extract the fitted values from an \code{\link{SVC_mle}} object. This is only possible if \code{cal.res} was set to \code{TRUE} in the control of the function call}
\usage{
\method{fitted}{SVC_mle}(object, ...)
}
\arguments{
\item{object}{\code{\link{SVC_mle}} object}

\item{...}{further arguments}
}
\value{
numeric, fitted values to given data
}
\description{
Extact Model Fitted Values

Method to extract the fitted values from an \code{\link{SVC_mle}} object. This is only possible if \code{cal.res} was set to \code{TRUE} in the control of the function call
}
