% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{own_dist}
\alias{own_dist}
\title{Computes (Cross-) Distances}
\usage{
own_dist(x, y = NULL, taper = NULL, ...)
}
\arguments{
\item{x}{(\code{matrix}) \cr Matrix containing locations}

\item{y}{(\code{NULL} or \code{matrix}) \cr If \code{NULL}, computes the
distances between \code{x}. Otherwise, computes cross-distances, i.e.,
pair-wise distances between rows of \code{x} and \code{y}.}

\item{taper}{(\code{NULL} or \code{numeric(1)}) \cr If \code{NULL}, all
distances are considered. Otherwise, only distances shorter than
\code{taper} are used. Hence the output will be a sparse matrix of type
\code{\link[spam]{spam}}.}

\item{...}{Further arguments for either \code{\link[stats]{dist}} or
\code{\link[spam]{nearest.dist}}.}
}
\value{
A \code{matrix} or \code{spam} object.
}
\description{
Computes (Cross-) Distances
}
