% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVC_mle.R
\name{predict.SVC_mle}
\alias{predict.SVC_mle}
\title{Prediction of SVC (and response variable)}
\usage{
\method{predict}{SVC_mle}(object, newlocs = NULL, newX = NULL,
  newW = NULL, compute.y.var = FALSE, ...)
}
\arguments{
\item{object}{output of \code{\link{SVC_mle}}}

\item{newlocs}{matrix of dimension n' x 2. These are the new locations the SVCs are predicted for. If \code{NULL}, the locations from the \code{SVC_mle} (i.e. \code{locs}) are considered.}

\item{newX}{optional matrix of dimension n' x pX. If provided, besides the predicted SVC, the function also returns the predicted response variable.}

\item{newW}{optional matrix of dimension n' x pW.}

\item{compute.y.var}{logical. If y will be estimated and \code{TRUE}, the standard deviation of each estimate will be computed.}

\item{...}{further arguments}
}
\value{
returns a data frame of n' rows and with columns
\itemize{
  \item \code{SVC_1, ..., SVC_p}, i.e. the predicted SVC at locations \code{newlocs}
  \item \code{y.pred}, if \code{newX} and \code{newW} are provided
  \item \code{y.var}, if \code{newX} and \code{newW} are provided and \code{compute.y.var} is set to \code{TRUE}.
  \item \code{loc_x, loc_y}, the locations of the predictions
}
}
\description{
Prediction of SVC (and response variable)
}
\seealso{
\code{\link{SVC_mle}}
}
