% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-SVC_mle.R
\name{print.SVC_mle}
\alias{print.SVC_mle}
\title{Print Method for \code{SVC_mle}}
\usage{
\method{print}{SVC_mle}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{\code{\link{SVC_mle}} object}

\item{digits}{(\code{numeric})
Number of digits to be plotted.}

\item{...}{further arguments}
}
\description{
Method to print an \code{\link{SVC_mle}} object.
}
\author{
Jakob Dambon
}
