% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-SVC_mle.R
\name{summary.SVC_mle}
\alias{summary.SVC_mle}
\title{Summary Method for \code{SVC_mle}}
\usage{
\method{summary}{SVC_mle}(object, ...)
}
\arguments{
\item{object}{\code{\link{SVC_mle}} object}

\item{...}{further arguments}
}
\value{
object of class \code{summary.SVC_mle} with summarized values of the MLE.
}
\description{
Method to construct a \code{summary.SVC_mle} object out of a \code{\link{SVC_mle}} object.
}
\seealso{
\code{\link{SVC_mle}}
}
\author{
Jakob Dambon
}
