% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVC_mle.R
\name{SVC_mle_control}
\alias{SVC_mle_control}
\title{Set Parameters for \code{SVC_mle}}
\usage{
SVC_mle_control(cov.name = c("exp"), tapering = NULL, cl = NULL,
  init = NULL, lower = NULL, upper = NULL, cal.res = FALSE, ...)
}
\arguments{
\item{cov.name}{name of the covariance function defining the covariance matrix of the GRF. Currently, only \code{"exp"} for the exponential is supported.}

\item{tapering}{if \code{NULL}, no tapering is applied. If a scalar is given, covariance tapering with this taper range is applied, for all GRF modelling the SVC.}

\item{cl}{cluster for parallelization. Currently not supported.}

\item{init}{numeric. Initial values for optimization procedure. The vector consists of p-times (alternating) scale and variance, the nugget variance and the p + p.fix mean effects}

\item{lower}{lower bound for optim, default \code{NULL} sets the lower bounds to 1e-6 for covariance parameters and \code{-Inf} for mean parameters.}

\item{upper}{upper bound for optim, default \code{NULL} sets the upper bounds to \code{Inf} for covariance and mean parameters.}

\item{cal.res}{logical. If \code{TRUE}, calculates the residuals after MLE.}

\item{...}{further parameters yet to be implemented}
}
\value{
A list with which \code{\link{SVC_mle}} can be controlled
}
\description{
Function to set up control parameters for \code{\link{SVC_mle}}
}
\examples{
control <- SVC_mle_control(init = rep(0.3, 10))
# or
control <- SVC_mle_control()
control$init <- rep(0.3, 10)
}
\seealso{
\code{\link{SVC_mle}}
}
