% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVC_mle.R
\name{coef.SVC_mle}
\alias{coef.SVC_mle}
\title{Extact Mean Effects}
\usage{
\method{coef}{SVC_mle}(object, ...)
}
\arguments{
\item{object}{\code{\link{SVC_mle}} object}

\item{...}{further arguments}
}
\value{
vector with mean effects, i.e. \eqn{\mu} from \code{\link{SVC_mle}}
}
\description{
Method to extract the mean effects from an \code{\link{SVC_mle}} object.
}
