% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_par.R
\name{cov_par}
\alias{cov_par}
\title{Extact Covariance Parameters

Function to extract the covariance parameters from an \code{\link{SVC_mle}} object.}
\usage{
cov_par(object, ...)
}
\arguments{
\item{object}{\code{\link{SVC_mle}} object}

\item{...}{further arguments}
}
\value{
vector with covariance parameters and attributes what kind
\itemize{
  \item \code{"GRF"}, charachter, describing the GRF used, see \code{\link{SVC_mle_control}}.
  \item \code{"tapering"}, either \code{NULL} if no tapering is applied of the taper range.
}
}
\description{
Extact Covariance Parameters

Function to extract the covariance parameters from an \code{\link{SVC_mle}} object.
}
\author{
Jakob Dambon
}
