% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals-SVC_mle.R
\name{residuals.SVC_mle}
\alias{residuals.SVC_mle}
\title{Extact Model Residuals}
\usage{
\method{residuals}{SVC_mle}(object, ...)
}
\arguments{
\item{object}{\code{\link{SVC_mle}} object}

\item{...}{further arguments}
}
\value{
numeric, residuals of model
}
\description{
Method to extract the residuals from an \code{\link{SVC_mle}} object. This is only possible if \code{save.fitted} was set to \code{TRUE} in the control of the function call
}
\author{
Jakob Dambon
}
