% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVC_mle.R
\name{SVC_mle.control}
\alias{SVC_mle.control}
\title{Set Parameters for \code{SVC_mle}}
\usage{
SVC_mle.control(cov.name = c("exp"), tapering = NULL, cl = NULL,
  scale = FALSE, init = NULL, ...)
}
\arguments{
\item{cov.name}{name of the covariance function defining the covariance matrix of the GRF. Currently, only \code{"exp"} for the exponential is supported.}

\item{tapering}{if \code{NULL}, no tapering is applied. If a scalar is given, covariance tapering with this taper range is applied.}

\item{cl}{cluster for parallelization. Currently not supported.}

\item{scale}{if \code{TRUE}, the covariates are being standardized. Currently not supported.}

\item{init}{numeric. Initial values for optimization procedure. The vector consists of p-times (alternating) scale and variance, the nugget variance and the p + p.fix mean effects}

\item{...}{further parameters yet to be implemented}
}
\value{
A list with which \code{\link{SVC_mle}} can be controlled
}
\description{
Function to set up control parameters for \code{\link{SVC_mle}}
}
\examples{
control <- SVC_mle.control()
control$init <- rep(0.3, 10) 
}
\seealso{
\code{\link{SVC_mle}}
}
