% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVC_mle.R
\name{SVC_mle}
\alias{SVC_mle}
\title{MLE of SVC model}
\usage{
SVC_mle(y, X, locs, W = NULL, control = SVC_mle.control(), ns = NULL,
  ...)
}
\arguments{
\item{y}{numeric response vector of dimension n.}

\item{X}{matrix of covariates of dimension n x pX. Intercept has to be added manually.}

\item{locs}{matrix of locations of dimension n X 2. May contain multiple observations at single location which (may) cause a permutation of \code{y}, \code{X}, \code{X.fixed} and \code{locs}.}

\item{W}{Optional matrix of covariates with fixed effects, i.e. non-SVC, of dimension n x pW}

\item{control}{list of control paramaters, usually given by \code{\link{SVC_mle.control}}}

\item{ns}{Do not use this argument.}

\item{...}{futher arguments for method}
}
\value{
Object of class \code{SVC_mle}
}
\description{
Calls MLE of the SVC model defined as:

\deqn{y(s) = X \mu + W \tilde \beta (s) + \epsilon(s)}

where:
\itemize{
  \item y is the response (vector of length n)
  \item X is the data matrix for the fixed effects covariates
  \item \eqn{\mu} is the vetor containing the fixed effects
  \item W is the data matrix for the SVCs represented by zero mean GRF
  \item \eqn{\tilde \beta} are the SVCs represented by zero mean GRF
  \item \eqn{\epsilon} is the nugget effect
}

The MLE is done by calling the function optim
}
\seealso{
\code{\link{SVC_mle}}, \code{\link{predict.SVC_mle}}
}
