% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVC_mle.R
\name{SVC_mle}
\alias{SVC_mle}
\title{MLE of SVC model}
\usage{
SVC_mle(y, X, locs, init, control = list(cov.name = "exp", tapering =
  NULL, cl = NULL, X.scale = FALSE), X.fixed = NULL, ns = NULL, ...)
}
\arguments{
\item{y}{numeric response vector of dimension n.}

\item{X}{matrix of covariates of dimension n x p. Intercept has to be added manually.}

\item{locs}{matrix of locations of dimension n X 2. May contain multiple observations at single location which (may) cause a permutation of \code{y}, \code{X}, \code{X.fixed} and \code{locs}.}

\item{init}{numeric. Initial values for optimization procedure. The vector consists of p-times (alternating) scale and variance, the nugget variance and the p + p.fix mean effects}

\item{control}{list of control paramaters, see Details}

\item{X.fixed}{Do not use this argument. Optional matrix of covariates with fixed effects, i.e. non-SVC, of dimension n x p.fixed}

\item{ns}{Do not use this argument.}

\item{...}{futher arguments for method}
}
\value{
Object of class \code{SVC_mle}
}
\description{
Calls MLE of the SVC model defined as:

\deqn{y(s) = x^{(1)}(s)\beta_1(s) + ... + x^{(p)}(s)\beta_p(s) + x^{(p+1)}\beta_{p+1} + ... + x^{(p+p.fix)}\beta_{p+p.fix} + \epsilon(s)}

where:
\itemize{
  \item p is the number of SVC
  \item \eqn{\beta_j} are distinctly defined Gaussian Random Fields, j = 1, ..., p
  \item p.fix is the number of fixed coefficients, which is optional
  \item \eqn{\epsilon} is the nugget effect
}

The MLE is done by calling the function optim
}
\details{
The \code{control} list has the following components
\itemize{
  \item \code{cov.name}, name of the covariance function defining the covariance matrix of the GRF. Currently, only \code{"exp"} for the exponential is supported.
  \item \code{tapering}, if \code{NULL}, no tapering is applied. If a scalar is given, covariance tapering with this taper range is applied.
  \item \code{cl}, cluster for parallelization. Currently not supported.
  \item \code{X.scale}, if \code{TRUE}, the covariates are being standardized. Currently not supported.
}
}
